FOS-QEMUX86_64: Support uboot loading 64bit linux kernel.
This commit includes patches to load kernel from MMC and
jump to it. The kernel and setup.bin is packed inside a
uboot fitimage and loaded from fat filesystem. The kernel
is stored as uncompressed vmlinuz format. The kernel
can be loaded 0x03000000 address and setup.bin to 0x090000.
Once loaded bootm command can be used to execute the
kernel.

TODO:
1) Support lzo compression on vmlinuz and uboot
2) Support loading fitimage directly from raw partition

Change-Id: Iac3d9b018a2da47314a28e5a31af4444cf8ba47e